Revert "Support Chrome Extension"
