+++ package.json
@@ -51,7 +51,6 @@
     "sweet-scroll": "^1.0.3",
     "vue": "^1.0.26",
     "vue-color": "0.0.7",
-    "vue-multiselect": "^1.1.4",
     "vue-resource": "^0.9.3",
     "vue-router": "^0.7.13",
     "vue-sharer": "^0.1.0",
