--- view/components/register/index.css
@@ -1,127 +1,0 @@
-@import '../../defs/import';
-
-@define-mixin _block-label {
-  display: block;
-  margin: 0 0 8px 0;
-  text-align: center;
-  font-size: $font-size-medium;
-  font-weight: bold;
-}
-
-.eg-register {
-  display: block;
-  position: relative;
-
-  .inner {
-    display: flex;
-    flex: 1 1 auto;
-    flex-direction: column;
-    align-content: flex-start;
-    align-items: center;
-    margin: 0 auto;
-    padding: 10px 0 10px 0;
-    transition: opacity .8s ease-in-out;
-    opacity: 1;
-  }
-
-  &.expand-transition {
-    padding: 20px 0 15px;
-    height: auto;
-    transition: opacity .8s ease-in-out;
-    opacity: 1;
-  }
-
-  &.expand-enter, &.expand-leave {
-    padding: 0;
-    height: 0;
-    overflow: hidden;
-    opacity: 0;
-  }
-
-  &.progress {
-    .overlay {
-      display: block;
-      z-index: $z-index-overlay;
-    }
-  }
-
-  .overlay {
-    display: none;
-    position: absolute;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    background-image: url('/static/img/loading_red.gif');
-    background-position: center center;
-    background-size: 40px auto;
-    background-repeat: no-repeat;
-    background-color: rgba(255, 255, 255, .6);
-  }
-
-  .groups {
-    label {
-      @mixin _block-label;
-      color: $color-blue-dark;
-    }
-
-    .group {
-      margin: 0 0 35px 0;
-
-      &.teams {
-        width: 400px;
-      }
-
-      &.name {
-        input {
-          box-shadow: 0 0 1px 1px rgba(0, 0, 0, .15);
-          border-radius: 2px;
-          padding: 7px 9px;
-          border: 0;
-          font-size: $font-size-medium;
-          width: 400px;
-        }
-      }
-    }
-  }
-
-  .messages {
-    margin: 10px 0 30px;
-    font-size: $font-size-small;
-    color: $color-gray;
-
-    label {
-      @mixin _block-label;
-      color: $color-blue-dark;
-    }
-
-    &.error {
-      color: $color-red-dark;
-      label {
-        color: $color-red-dark;
-      }
-    }
-
-    p {
-      margin: 0 0 5px;
-    }
-  }
-
-  .buttons {
-    margin-top: 10px;
-    text-align: center;
-
-    button {
-      padding: 12px 60px;
-      border-radius: 12px;
-      background-color: color($color-red a(.50));
-      background-image: none !important;
-      color: white;
-      font-weight: bold;
-
-      &:hover, &:focus {
-        background-color: color($color-red a(.75));
-      }
-    }
-  }
-}
