--- view/components/register/index.html
@@ -1,43 +1,0 @@
-<div class="v-cloak eg-register" :class="{ 'progress': progress }">
-  <div class="overlay"></div>
-  <div class="inner">
-    <div class="groups">
-      <div class="group teams">
-        <label>登録するチーム</label>
-        <multiselect
-          :selected="selected"
-          :options="teams"
-          :close-on-select="true"
-          :show-labels="false"
-          :loading="progressTeams"
-          @update="updateSelectedTeam"
-          label="name"
-          placeholder="登録するチームを選んで下さい"
-        ></multiselect>
-      </div>
-      <div class="group name">
-        <label>絵文字の名前</label>
-        <input type="text" placeholder="絵文字の名前を入れて下さい" v-model="text">
-      </div>
-    </div>
-
-    <div class="messages error" v-if="visibleErrors && hasErrorMessages">
-      <label>入力を確認してください</label>
-      <p v-for="errorMessage in errorMessages">{{ errorMessage }}</p>
-    </div>
-
-    <div class="messages" v-if="result.contents">
-      <label>絵文字を登録しました！</label>
-      <p>{{ result.contents }}</p>
-    </div>
-
-    <div class="messages error" v-if="result.err">
-      <label>絵文字の登録に失敗しました</label>
-      <p>{{ result.err }}</p>
-    </div>
-
-    <div class="buttons">
-      <button type="button" class="pure-button" @click="register">登録する&#9834;</button>
-    </div>
-  </div>
-</div>
