--- view/components/share/index.html
@@ -1,7 +1,0 @@
-<div class="v-cloak eg-share" :class="{ 'progress': progress }">
-  <div class="inner">
-    <input type="button" class="twitter sharer button" data-sharer="twitter" data-title="絵文字ジェネレーター使って絵文字を生成しました&#9834; {{shortenUrl}} #絵文字ジェネレーター" title="Twitter でシェアする" v-sharer>
-    <input type="button" class="facebook sharer button" data-sharer="facebook" :data-url="shortenUrl" title="Facebook でシェアする" v-sharer>
-    <input type="button" class="google sharer button" data-sharer="googleplus" :data-url="shortenUrl" title="Google+ でシェアする" v-sharer>
-  </div>
-</div>
