+++ view/parts/result/index.css
@@ -15,6 +15,7 @@
   }
 
   &.expand-transition {
+    overflow: hidden;
     height: auto;
     transition: opacity .8s ease-in-out;
     opacity: 1;
@@ -23,7 +24,6 @@
   &.expand-enter, &.expand-leave {
     padding: 0;
     height: 0;
-    overflow: hidden;
     opacity: 0;
   }
 
@@ -129,15 +129,10 @@
       display: flex;
 
       > div {
-        margin-left: 40px;
         padding: 0;
         height: $_height;
         line-height: $_height;
         font-size: $font-size-medium;
-
-        &:first-child {
-          margin-left: 0;
-        }
 
         a {
           display: block;
@@ -157,18 +152,8 @@
           }
         }
 
-        &.register {
-          padding: 0 0 0 34px;
-          color: $color-red-dark;
-          height: 100%;
-          background-image: url('/static/img/register.png');
-          background-repeat: no-repeat;
-          background-position: 0 center;
-          background-size: 22px auto;
-          cursor: pointer;
-        }
-
         &.share {
+          margin: 0 0 0 40px;
           padding: 0 0 0 34px;
           color: $color-yellow-dark;
           background-image: url('/static/img/share.png');
@@ -181,4 +166,54 @@
     }
   }
 
+  > .share {
+    display: block;
+
+    .inner {
+      display: flex;
+      flex: 1 1 auto;
+      flex-direction: row;
+      justify-content: center;
+      align-items: center;
+      transition: opacity .8s ease-in-out;
+      opacity: 1;
+
+      input, a {
+        @mixin icon-40px;
+        margin: 0 0 0 14px;
+
+        &.twitter {
+          @mixin icon-twitter;
+          margin-left: 0;
+        }
+        &.facebook { @mixin icon-facebook }
+        &.google { @mixin icon-google }
+      }
+    }
+
+    &.expand-transition {
+      padding: 20px 0 15px;
+      overflow: hidden;
+      height: auto;
+      transition: opacity .8s ease-in-out;
+      opacity: 1;
+    }
+
+    &.expand-enter, &.expand-leave {
+      padding: 0;
+      height: 0;
+      opacity: 0;
+    }
+
+    &.progress {
+      background-image: url('/static/img/loading.gif');
+      background-position: center 10px;
+      background-size: 40px auto;
+      background-repeat: no-repeat;
+
+      .inner {
+        opacity: 0;
+      }
+    }
+  }
 }
