+++ view/parts/result/index.html
@@ -33,25 +33,17 @@
       <div class="download">
         <a :href="emojiDownloadUrl">ダウンロード</a>
       </div>
-      <div class="register" @click="toggleRegister" v-show="hasChromeExtension">
-        登録する
-      </div>
-      <div class="share" @click="toggleShare">
+      <div class="share" v-on:click="toggleShare">
         シェアする
       </div>
     </div>
   </div>
 
-  <eg-share
-    v-show="visibleShare"
-    transition="expand"
-    :visible="visibleShare"
-  ></eg-share>
-
-  <eg-register
-    v-show="visibleRegister"
-    transition="expand"
-    :visible="visibleRegister"
-    :emoji-url="emojiUrl"
-  ></eg-share>
+  <div class="share" :class="{ 'progress': progress }" v-show="visibleShare" transition="expand">
+    <div class="inner">
+      <input type="button" class="twitter sharer button" data-sharer="twitter" data-title="絵文字ジェネレーター使って絵文字を生成しました&#9834; {{shortenUrl}} #絵文字ジェネレーター" title="Twitter でシェアする" v-eg-sharer>
+      <input type="button" class="facebook sharer button" data-sharer="facebook" :data-url="shortenUrl" title="Facebook でシェアする" v-eg-sharer>
+      <input type="button" class="google sharer button" data-sharer="googleplus" :data-url="shortenUrl" title="Google+ でシェアする" v-eg-sharer>
+    </div>
+  </div>
 </div>
